+++ packages/core/src/tracing/clsHooked/context.js
@@ -15,7 +15,7 @@
 const util = require('util');
 const assert = require('assert');
 // @ts-ignore
-const wrapEmitter = require('emitter-listener');
+const wrapEmitter = require('./emitter-listener');
 const async_hooks = require('async_hooks');
 const unset = require('./unset');